requirements.txt
discord.py==2.3.2
openai==1.3.8
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
notion-client==2.2.1
Pillow==10.1.0
mistralai==0.4.2
aiohttp==3.9.1
google-cloud-aiplatform==1.49.0
google-cloud-secret-manager==2.20.2
PyPDF2==3.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyyaml==6.0.1